requirements.txt
# Core LLM
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0

# Schema & validation
pydantic>=2.0

# Document parsing
PyMuPDF>=1.23.0
python-docx>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Data analysis
pandas>=2.0.0

# Config
python-dotenv>=1.0.0

# API serving (optional)
# fastapi>=0.110.0
# uvicorn>=0.29.0
